pyproject.toml
[tool.poetry]
name = "openrouter-examples-python"
version = "0.1.0"
description = ""
authors = ["OpenRouter <support@openrouter.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
langchain = "^0.0.335"
litellm = "^0.14.1"
openai = "^1.21.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
